format.py file referencing is now more generic
Since format.py used to exist within a "styleguide" directory in WPILib, it was hardcoded to check whether it was called from within that directory and added "../" to file paths if so. Since this repository is called "styleguide", format.py erroneously adds "../" to all file paths when run on this repository.

This commit modifies format.py to search for the ".git" folder in the repo's root directory and references files based on its location.