includeorder.py now retains comment prefix in commented-out includes
A warning is generated for commented-out includes. The contents of the
include lists were changed from strings to regex matches so the contents
could be sorted by the name group. Since sets can't be used to deduplicate
the regex matches before sorting and sorting the regex matches produces a
list, a custom deduplication function was used instead after the regex
matches were converted to strings.