Fix file change detection
Previously, a file was considered changed and rewritten if any file in a
pipeline changed it. Now, it's only considered changed if the input to
the pipeline doesn't equal the output. This allows tasks later in the
pipeline to fix the output of previous tasks.

To make this work across more scenarios, clang-format had to be changed
to run on a per-file basis instead of batched. This results in an
increase in processing time due to process creation overhead.