Bump copyright year if a file has uncommitted changes
This obviates needing to rerun wpiformat after a commit if the file had
a previous copyright year and the new commit is done in a later year.

If the file is being amended into a commit with a date from a previous
year such that the old year is kept, the year set by uncommitted changes
will be incorrect and will require rerunning wpiformat as before. This
was already necessary so this patch is strictly an improvement.