Lock regex version to 2021.08.21 (#211)
The published package for 2021.08.27 hangs on exit in multiprocessing
and threading waiting on I/O. I was able to reproduce this locally with
2021.08.27, and 2021.08.21 didn't exhibit the issue.

I also tried installing local builds of the regex git repo for the
commits on 8/21 and 8/27; they both worked fine. The issue appears to be
unique to the published package for 2021.08.27.